factorize code in graphService
