+++ src/tests/stores/autocalsolResultExample.json
@@ -23,11 +23,11 @@
     ["10:00:00", 374],
     ["11:00:00", 95],
     ["12:00:00", 243],
-    ["13:00:00", 505],
-    ["14:00:00", 648],
-    ["15:00:00", 464],
-    ["16:00:00", 251],
-    ["17:00:00", 227],
+    ["13:00:00", 750],
+    ["14:00:00", 900],
+    ["15:00:00", 950],
+    ["16:00:00", 1000],
+    ["17:00:00", 700],
     ["18:00:00", 84],
     ["19:00:00", 14],
     ["20:00:00", 0],
